MOVE: 에러 관련 코드 이동
