+++ backend/src/main/java/team03/secondhand/domain/member/MemberControllerAdvice.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import team03.secondhand.domain.BaseResponse;
 import team03.secondhand.domain.StatusCode;
-import team03.secondhand.domain.member.error.MemberError;
+import team03.secondhand.error.MemberError;
 
 @RestControllerAdvice(basePackages = "team03.secondhand.domain.member")
 public class MemberControllerAdvice {
